Make the gem compatible with the rails asset pipeline.
In the project I'm using this one we are using the javascript on the
frontend to validate user input as they type. Right now the file is
being manually copied over. It would be preferable to just be able to
require it, since it is already present in this gem, which we are
dependent on.